Update hooks structure requirements
The current hooks structure has package.json within the lib folder. This causes strapi to fail as it is looking for the package.json file in the hook root folder and not within lib.

This change moves the secondary files back into the hook root leaving only index.js in the lib folder.